+++ CHANGELOG.md
@@ -17,6 +17,7 @@ for a function that will run when accepting a new connection.
 ### Bug fixes
 
 - Isolate copy paste inference per readline session ([#197](https://github.com/anmonteiro/lumo/issues/197)).
+- Socket REPL fails to isolate NS ([#158](https://github.com/anmonteiro/lumo/issues/158)).
 
 ## [1.6.0](https://github.com/anmonteiro/lumo/compare/1.5.0...1.6.0) (2017-06-30)