+++ CHANGELOG.md
@@ -14,7 +14,7 @@ for a function that will run when accepting a new connection.
 
 - Use Paredit.js to calculate indentation for multiline forms ([#193](https://github.com/anmonteiro/lumo/issues/193)).
 - Upgrade Google Closure Compiler to v20170626.
-- Upgrade Node.js to version 8.2.0.
+- Upgrade Node.js to version 8.2.1.
 
 ### Bug fixes